solve some sms issues
