+++ .env
@@ -1,0 +1,5 @@
+api_key=04c61df811b0529200d6186b16e7948b2b5a73bf44bf13d8634b80dd2dc4bac8
+
+username=sandbox
+
+alphanumeric=AUTHMSG
